updated bellsensorstage2.cpp
Updated firmware (untested): need to wait 250ms after powering up the gyro, and discard whatever is in the fifo so we don't send pre-buffered (and probably bogus) packets.
Increase delay time to sleeping to 15 minutes, since we cannot hope to get good readings on the first pull off.

TODO: need to better tune the accelerometer wake-up so that it is more sensitive when the bell is pulled off the balance.

TODO: Don't send accelerometer data when the sensor is in motion.